+++ documentation/conf.py
@@ -25,3 +25,4 @@ language = 'en'
 
 html_theme = 'sphinx_rtd_theme'
 html_static_path = ['_static']
+html_css_files = ['css/extra.css']