+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Added the ability to organize taxonomy terms in a hierarchy
 - Fixed unit tests
 
 ## [0.1.1] - 2023-06-29